guest-mode local data save ui changes bug fix

+++ domain/src/main/java/com/stefanenko/coinbase/domain/entity/CurrencyMarketInfo.kt
@@ -10,5 +10,5 @@ data class CurrencyMarketInfo(
     val action: String,
 
     @SerializedName("price")
-    val price: Double
+    val price: Float
 )
